+++ Docker/bdg-sequila/bin/featureCounts.sh
@@ -17,7 +17,7 @@ if case ${sparkParams} in *"${substr}"*) true;; *) false;; esac; then
 
 outfile=`echo $appParams | sed -n "s/^.*-o \([^ ]*\) .*$/\1/p"`
 
-#echo "Checking output directory " $outfile
+echo "Output " $outfile
 
 #if [ -e "$outfile" ]
 #then
@@ -42,6 +42,9 @@ echo "Arguments passed to Apache Spark: $sparkParams"
 echo -e "\n"
 spark-submit -v ${sparkParams} --class org.biodatageeks.apps.FeatureCounts /tmp/bdg-toolset/bdg-sequila-assembly-${BDG_VERSION}.jar  $appParams
 
-mv $outfile/p*  $outfile_tmp
+if [ -e "$outfile" ]
+then
+mv ${outfile}/part*  ${outfile}_tmp
 rm -rf $outfile
-mv $outfile_tmp $outfile
+mv ${outfile}_tmp $outfile
+fi
